# generated/src/main/java/org/openapitools/model/PaymentOrderInitiationResponse.java
package org.openapitools.model;

import java.net.URI;
import java.util.Objects;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonValue;
import org.springframework.lang.Nullable;
import org.openapitools.jackson.nullable.JsonNullable;
import java.time.OffsetDateTime;
import javax.validation.Valid;
import javax.validation.constraints.*;
import io.swagger.v3.oas.annotations.media.Schema;


import java.util.*;
import javax.annotation.Generated;

/**
 * PaymentOrderInitiationResponse
 */

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2025-11-18T22:49:19.373248300-05:00[America/Guayaquil]", comments = "Generator version: 7.17.0")
public class PaymentOrderInitiationResponse {

  private @Nullable String paymentOrderId;

  /**
   * Gets or Sets status
   */
  public enum StatusEnum {
    INITIATED("INITIATED"),
    
    VALIDATED("VALIDATED"),
    
    PENDING("PENDING"),
    
    EXECUTED("EXECUTED"),
    
    REJECTED("REJECTED");

    private final String value;

    StatusEnum(String value) {
      this.value = value;
    }

    @JsonValue
    public String getValue() {
      return value;
    }

    @Override
    public String toString() {
      return String.valueOf(value);
    }

    @JsonCreator
    public static StatusEnum fromValue(String value) {
      for (StatusEnum b : StatusEnum.values()) {
        if (b.value.equals(value)) {
          return b;
        }
      }
      throw new IllegalArgumentException("Unexpected value '" + value + "'");
    }
  }

  private @Nullable StatusEnum status;

  public PaymentOrderInitiationResponse paymentOrderId(@Nullable String paymentOrderId) {
    this.paymentOrderId = paymentOrderId;
    return this;
  }

  /**
   * Get paymentOrderId
   * @return paymentOrderId
   */
  
  @Schema(name = "paymentOrderId", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
  @JsonProperty("paymentOrderId")
  public @Nullable String getPaymentOrderId() {
    return paymentOrderId;
  }

  public void setPaymentOrderId(@Nullable String paymentOrderId) {
    this.paymentOrderId = paymentOrderId;
  }

  public PaymentOrderInitiationResponse status(@Nullable StatusEnum status) {
    this.status = status;
    return this;
  }

  /**
   * Get status
   * @return status
   */
  
  @Schema(name = "status", requiredMode = Schema.RequiredMode.NOT_REQUIRED)
  @JsonProperty("status")
  public @Nullable StatusEnum getStatus() {
    return status;
  }

  public void setStatus(@Nullable StatusEnum status) {
    this.status = status;
  }

  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    PaymentOrderInitiationResponse paymentOrderInitiationResponse = (PaymentOrderInitiationResponse) o;
    return Objects.equals(this.paymentOrderId, paymentOrderInitiationResponse.paymentOrderId) &&
        Objects.equals(this.status, paymentOrderInitiationResponse.status);
  }

  @Override
  public int hashCode() {
    return Objects.hash(paymentOrderId, status);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class PaymentOrderInitiationResponse {\n");
    sb.append("    paymentOrderId: ").append(toIndentedString(paymentOrderId)).append("\n");
    sb.append("    status: ").append(toIndentedString(status)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}